Check PagedAttention sanitizer sizes

+++ onnxruntime/contrib_ops/cuda/bert/paged_attention_impl.h
@@ -42,7 +42,7 @@ Status LaunchSanitizeSequenceLengths(int32_t* sanitized_cumulative_seqlens_q,
                                      cudaStream_t stream);
 
 Status LaunchSanitizeBlockTable(const int32_t* block_table, int32_t* sanitized_block_table,
-                                int element_count, int num_blocks, cudaStream_t stream);
+                                size_t element_count, int num_blocks, cudaStream_t stream);
 // Produces per-batch KV lengths seqlens_kv[i] = past_seqlens[i] + (cumulative_seqlens_q[i+1] -
 // cumulative_seqlens_q[i]) for the cuDNN paged SDPA backend's padding-mask input. Deriving the
 // query count from cumulative_seqlens_q keeps the kernel correct if a caller ever routes a
